Ugly Number
29th Problem - Ugly Number
Description: An ugly number is a positive integer whose only prime factors are 2, 3, and 5. The task is to check if a given integer n is an ugly number.

Approach: Check if n is positive; if not, return false. Then, repeatedly divide n by 2, 3, and 5 until it cannot be divided anymore. If the resulting value is 1, return true (ugly number); otherwise, return false.