Same Tree
40th problem -Same Tree
Description:
The function isSameTree checks if two binary trees are identical by comparing their structure and node values. It returns true if the trees are identical and false otherwise.

Approach:

If both nodes are null, they are identical, so return true.
If only one of the nodes is null, they differ, so return false.
If the values of p and q are equal, recursively check their left and right subtrees. If both are identical, return true.
Otherwise, return false.