Number steps to Reduce a Number to Zero
42nd Problem -  Number steps to Reduce a Number to Zero
Description: Given an integer num, the task is to return the number of steps required to reduce it to zero. In each step, if the number is even, divide it by 2; if it's odd, subtract 1.

Approach: Use a while loop to continuously check if the number is even or odd, applying the respective operation until num reaches zero, and count each operation to determine the total steps.