Is Subsequence
46th Problem - Is Subsequence
Description: Given two strings s and t, determine if s is a subsequence of t by ensuring all characters of s appear in t in the same relative order.

Approach: Use two pointers (sPointer for s and tPointer for t) to traverse t. Move sPointer only when matching characters are found, and return true if sPointer reaches the end of s.