Minimum Array End
60th Problem -  Minimum Array End
Description: Given integers n and x, construct an increasing array nums of size n such that the bitwise AND of all elements equals x. Return the minimum possible value of the last element of nums.

Approach: Start with ans = x. For each additional element, increment ans by 1 and perform a bitwise OR with x, ensuring that the bitwise AND with the initial element remains x. Repeat this for n-1 times to obtain the smallest possible nums[n-1].