Fizz Buzz
64th Problem -  Fizz Buzz

Description: Iterates through 1 to n, checks if each number is divisible by 3 and/or 5, and adds the corresponding "Fizz", "Buzz", or "FizzBuzz" string to the answer list. If not divisible by 3 or 5, add the number as a string.

Approach: Iterate through numbers, check divisibility by 3 and 5, and add an appropriate string to the list.